API/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2.1
django-cors-headers==3.14.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.3
mysqlclient==2.1.1
psycopg==3.1.9
psycopg2-binary==2.9.6
pycodestyle==2.10.0
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
